pyproject.toml
[project]
name = "ai-resume-job-matcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nltk>=3.9.2",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "python-docx>=1.2.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
    "spacy>=3.8.7",
    "streamlit>=1.50.0",
]
